Prepared Cell&Disc for rework - EH
Current use of Cell and Disc is redundant. One needs to be removed, and simply be replaced by the other. Thus I have made efforts to have Disc remain, But have yet to remove Cell entirely. Separately I have established a simple display function for the TPC4F as well as added functionality between the display and adding new pieces.