Display design
Made a Display for single player, the idea is that the user selects a purple button over a row and it changes the bottom most white button to the players color.  Right after the AI would make its move and drop its color.  Let me know if you all are okay with how it looks