Further SinglePlayerAi Improvements - EH
all but a few small errors regarding the Ai and knowing when to place a piece to block/win with a 4 in a row, and when to avoid a space to insure the player can't take advantage of that square. Of which should be completely sorted out during my next session.