Ai "Done", fixed TP win checking - EH
The Single Player AI should now be fully "Functional" as I have done about 10+ tests since my final changes, and have not encountered any errors. As well as I fixed the two player win checker: due to it only triggering if it was the last piece placed on the end of 3 others. It wouldn't register when placing the 4th piece in between the other 3.